+++ src/app/eform/eform.module.ts
@@ -10,13 +10,14 @@ import {
 import {CollapsibleWidget} from '../widgets/collapsable/collapsable.widget';
 import {FormsModule} from '@angular/forms';
 import {BrowserModule} from '@angular/platform-browser';
+import {UtilsService} from '../services/utils.service';
 
 @NgModule({
     imports: [  BrowserModule,FormsModule,AccordionModule,SchemaFormModule],
     exports:[EformComponent,PreviewComponent],
     declarations: [EformComponent,PreviewComponent,CollapsibleWidget],
   entryComponents: [CollapsibleWidget],
-  providers: [{provide: WidgetRegistry, useClass: DefaultWidgetRegistry}]
+  providers: [{provide: WidgetRegistry, useClass: DefaultWidgetRegistry},UtilsService]
 })
 export class EformModule {
   constructor(widgetRegistry: WidgetRegistry) {
