+++ src/app/services/utils.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import {Observable} from 'rxjs';
 
 @Injectable()
 export class UtilsService {
@@ -7,4 +8,20 @@ export class UtilsService {
 
     }
 
+  parseFileToObject(file: any) :Observable<any> {
+    return Observable.create((observer)=>{
+
+      var reader = new FileReader();
+      reader.addEventListener('load', (result) => {
+        observer.next( this.parseJsonFile(result));
+      });
+      reader.readAsText(file);
+    })
+  }
+
+
+  private parseJsonFile(result) {
+    var target: any = result.target.result;
+    return JSON.parse(target);
+  }
 }
